# DotNetServer/src/IO.Swagger/Models/PropertyType.cs
/*
 * Aspose.Pdf RestApi
 *
 * Demo 1
 *
 * OpenAPI spec version: v1
 * Contact: Evgeny.Pirogov@aspose.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{
  
          /// <summary>
          /// Gets or Sets PropertyType
          /// </summary>
          [JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
          public enum PropertyType
          {
              /// <summary>
              /// Enum NUMBER_0 for 0
              /// </summary>
              [EnumMember(Value = "0")]
              NUMBER_0 = 0,
              /// <summary>
              /// Enum NUMBER_1 for 1
              /// </summary>
              [EnumMember(Value = "1")]
              NUMBER_1 = 1,
              /// <summary>
              /// Enum NUMBER_2 for 2
              /// </summary>
              [EnumMember(Value = "2")]
              NUMBER_2 = 2,
              /// <summary>
              /// Enum NUMBER_3 for 3
              /// </summary>
              [EnumMember(Value = "3")]
              NUMBER_3 = 3,
              /// <summary>
              /// Enum NUMBER_4 for 4
              /// </summary>
              [EnumMember(Value = "4")]
              NUMBER_4 = 4,
              /// <summary>
              /// Enum NUMBER_5 for 5
              /// </summary>
              [EnumMember(Value = "5")]
              NUMBER_5 = 5,
              /// <summary>
              /// Enum NUMBER_6 for 6
              /// </summary>
              [EnumMember(Value = "6")]
              NUMBER_6 = 6,
              /// <summary>
              /// Enum NUMBER_7 for 7
              /// </summary>
              [EnumMember(Value = "7")]
              NUMBER_7 = 7,
              /// <summary>
              /// Enum NUMBER_8 for 8
              /// </summary>
              [EnumMember(Value = "8")]
              NUMBER_8 = 8          }
}