# DotNetServer/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Aspose.Pdf RestApi",
    "description" : "Demo 1",
    "termsOfService" : "https://asopse.com",
    "contact" : {
      "name" : "Evgeny Pirogov",
      "email" : "Evgeny.Pirogov@aspose.com"
    },
    "license" : {
      "name" : "Aspose.Pdf license"
    },
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
    "/pdf/webapi/status/{id}" : {
      "post" : {
        "tags" : [ "Api" ],
        "summary" : "Retrieves the status of a processing.",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The ID of the processing to check the status of.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a FileResponse object with the status information.",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FileResponse"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FileResponse"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FileResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "If the file does not exist or if there was an error processing the request."
          },
          "500" : {
            "description" : "If there was a server error."
          }
        }
      }
    },
    "/pdf/webapi/download/{id}" : {
      "get" : {
        "tags" : [ "Api" ],
        "summary" : "Downloads a file.",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The id of the processing to download.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "file",
          "in" : "query",
          "description" : "The name of the file to download.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns the file to download."
          },
          "400" : {
            "description" : "If the file does not exist."
          },
          "500" : {
            "description" : "If there was a server error."
          }
        }
      }
    },
    "/pdf/webapi/download-stream/{id}" : {
      "post" : {
        "tags" : [ "Api" ],
        "summary" : "Downloads a file as a stream.",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The ID of the processing to download.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "fileName",
          "in" : "query",
          "description" : "The name of the file to download.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns the file stream and metadata.",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FileResponse"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FileResponse"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FileResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "If the file does not exist or if there was an error processing the request."
          },
          "500" : {
            "description" : "If there was a server error."
          }
        }
      }
    },
    "/pdf/webapi/remove/{id}" : {
      "post" : {
        "tags" : [ "Api" ],
        "summary" : "Removes all files from processing.",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The ID of the processing to remove.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a FileResponse object indicating that all files was successfully removed.",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FileResponse"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FileResponse"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FileResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "If the file does not exist or if there was an error processing the request."
          },
          "500" : {
            "description" : "If there was a server error."
          }
        }
      }
    },
...
    "/pdf/webapi/gif" : {
      "post" : {
        "tags" : [ "Api" ],
        "summary" : "Creates a GIF from the provided images.",
        "description" : "This endpoint is used to create a GIF from the provided images.\r\nIf the creation is started, the response will contain a success status and an optional message.\r\nIf the creation fails, the response will contain an error status and a message indicating the reason for the failure.",
        "parameters" : [ {
          "name" : "keepImagesAsIs",
          "in" : "query",
          "description" : "Indicates whether the images should be kept as is.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "resizeToImage",
          "in" : "query",
          "description" : "The index of the image to resize the other images to.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "streched",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "width",
          "in" : "query",
          "description" : "The width of the GIF.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "height",
          "in" : "query",
          "description" : "The height of the GIF.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "frameTimes",
          "in" : "query",
          "description" : "The times for each frame in the GIF.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The resource was successfully added to processing.",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FileResponse"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FileResponse"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FileResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "The request was invalid or the resource could not be locked."
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "DocProperty" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string",
            "nullable" : true
          },
          "name" : {
            "type" : "string",
            "nullable" : true
          },
          "value" : {
            "nullable" : true
          },
          "type" : {
            "$ref" : "#/components/schemas/PropertyType"
          },
          "readOnly" : {
            "type" : "boolean"
          }
        },
        "additionalProperties" : false,
        "description" : "The same fields as in DocumentProperty",
        "example" : {
          "name" : "name",
          "readOnly" : true,
          "type" : 1,
          "value" : "",
          "key" : "key"
        }
      },
      "FileProcessingErrorCode" : {
        "type" : "integer",
        "format" : "int32",
        "enum" : [ 0, 10000, 10001, 10002 ]
      },
      "FileResponse" : {
        "type" : "object",
        "properties" : {
          "statusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "status" : {
            "type" : "string",
            "nullable" : true
          },
          "text" : {
            "type" : "string",
            "nullable" : true
          },
          "locked" : {
            "type" : "boolean"
          },
          "updated" : {
            "type" : "string",
            "description" : "Last updated",
            "format" : "date-time",
            "nullable" : true
          },
          "sharedFiles" : {
            "type" : "array",
            "nullable" : true,
            "items" : {
              "$ref" : "#/components/schemas/FileSafeResult"
            }
          },
          "files" : {
            "type" : "array",
            "nullable" : true,
            "items" : {
              "type" : "string"
            }
          },
          "fileProcessingErrorCode" : {
            "$ref" : "#/components/schemas/FileProcessingErrorCode"
          },
          "fileCount" : {
            "type" : "integer",
            "description" : "Get count of files",
            "format" : "int32",
            "readOnly" : true
          },
          "fileName" : {
            "type" : "string",
            "nullable" : true
          },
          "folderName" : {
            "type" : "string",
            "nullable" : true
          },
          "resultData" : {
            "nullable" : true
          },
          "appInstanceId" : {
            "type" : "string",
            "description" : "Where is processing located",
            "nullable" : true
          }
        },
        "additionalProperties" : false,
        "example" : {
          "fileName" : "fileName",
          "resultData" : "",
          "fileProcessingErrorCode" : 5,
          "fileCount" : 5,
          "appInstanceId" : "appInstanceId",
          "files" : [ "files", "files" ],
          "text" : "text",
          "folderName" : "folderName",
          "locked" : true,
          "updated" : "2000-01-23T04:56:07.000+00:00",
          "sharedFiles" : [ {
            "fileName" : "fileName",
            "uploadFileName" : "uploadFileName",
            "customProp" : [ null, null ],
            "id" : "id",
            "fileLength" : 6,
            "isSuccess" : true,
            "idUpload" : "idUpload",
            "builtProp" : [ {
              "name" : "name",
              "readOnly" : true,
              "type" : 1,
              "value" : "",
              "key" : "key"
            }, {
              "name" : "name",
              "readOnly" : true,
              "type" : 1,
              "value" : "",
              "key" : "key"
            } ]
          }, {
            "fileName" : "fileName",
            "uploadFileName" : "uploadFileName",
            "customProp" : [ null, null ],
            "id" : "id",
            "fileLength" : 6,
            "isSuccess" : true,
            "idUpload" : "idUpload",
            "builtProp" : [ {
              "name" : "name",
              "readOnly" : true,
              "type" : 1,
              "value" : "",
              "key" : "key"
            }, {
              "name" : "name",
              "readOnly" : true,
              "type" : 1,
              "value" : "",
              "key" : "key"
            } ]
          } ],
          "statusCode" : 0,
          "status" : "status"
        }
      },
      "FileSafeResult" : {
        "type" : "object",
        "properties" : {
          "isSuccess" : {
            "type" : "boolean",
            "description" : "Is result success?"
          },
          "id" : {
            "type" : "string",
            "description" : "Upload id.",
            "nullable" : true
          },
          "fileName" : {
            "type" : "string",
            "description" : "File name.",
            "nullable" : true
          },
          "uploadFileName" : {
            "type" : "string",
            "description" : "Gets or sets the name of the upload file.",
            "nullable" : true
          },
          "idUpload" : {
            "type" : "string",
            "description" : "idUpload from request.",
            "nullable" : true
          },
          "fileLength" : {
            "type" : "integer",
            "description" : "File length.",
            "format" : "int64",
            "nullable" : true,
            "readOnly" : true
          },
          "builtProp" : {
            "type" : "array",
            "nullable" : true,
            "items" : {
              "$ref" : "#/components/schemas/DocProperty"
            }
          },
          "customProp" : {
            "type" : "array",
            "nullable" : true,
            "items" : {
              "$ref" : "#/components/schemas/DocProperty"
            }
          }
        },
        "additionalProperties" : false,
        "description" : "File processing result.",
        "example" : {
          "fileName" : "fileName",
          "uploadFileName" : "uploadFileName",
          "customProp" : [ null, null ],
          "id" : "id",
          "fileLength" : 6,
          "isSuccess" : true,
          "idUpload" : "idUpload",
          "builtProp" : [ {
            "name" : "name",
            "readOnly" : true,
            "type" : 1,
            "value" : "",
            "key" : "key"
          }, {
            "name" : "name",
            "readOnly" : true,
            "type" : 1,
            "value" : "",
            "key" : "key"
          } ]
        }
      },
      "PropertyType" : {
        "type" : "integer",
        "format" : "int32",
        "enum" : [ 0, 1, 2, 3, 4, 5, 6, 7, 8 ]
      },
...
    }
  }
}